Fix intermittent issues with Home Assistant UI
The HA UI is out of sync with the actual target settings most of the time.
This happens because flux_led always reports the actual current values of the
device by querying the device directly. After changing the brightness in
the UI, HA tries to update the UI directly after issuing the brightness or
color change, but the device may fade to the target brightness/color instead of
directly applying it. When HA gets the property for color & brightness from
flux_led in this case, directly after triggering the brightness/color change,
it can happen that the device reports some kind of intermediate value because
the device is currently in the process of fading to the target brightness/color.

This means that the HA UI will immediately jump back to some intermediate value
right after trying to change the color or brightness, which is obviously totally
confusing. This will recover the next time HA updates the state from flux_led
and the device finished fading to the target brightness/color.

To address this issue, we have to cache the target value of the device and
report these when HA asks for the color and brightness. The internal cache
will then be updated when the next update_state call happens. For HA, this is
usually every 30 seconds or so, so the device will have usually reached the
target color/brightness by then and actual device state and the cached values
should already be in sync. In case some other actor changed the settings in the
meantime, the cache and thus HA UI will be updated with the actual device values.